Rename directory in truffle-reporters, adjust the way Truffle accesses the migration reporter, and adjust the files requiring the reporter
