+++ packages/truffle-deployer/test/errors.js
@@ -1,7 +1,7 @@
 const ganache = require("ganache-core");
 const Web3 = require("web3");
 const assert = require("assert");
-const Reporter = require("truffle-reporters").migrationsV5;
+const Reporter = require("truffle-reporters").migration;
 const EventEmitter = require("events");
 
 const Deployer = require("../index");
@@ -70,8 +70,8 @@ describe("Error cases", function() {
       network_id: networkId,
       provider: provider,
       logger: {
-        log: val => {},
-        error: val => {}
+        log: () => {},
+        error: () => {}
       }
     };
     deployer = new Deployer(options);
