+++ packages/truffle-reporters/index.js
@@ -1,4 +1,4 @@
 module.exports = {
-  migrationsV5: require("./reporters/migrations-V5/reporter"),
-  compilationReporter: require("./compilation")
+  migration: require("./migration"),
+  compilation: require("./compilation")
 };
